requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
fastavro==1.10.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
idna==3.10
protobuf==5.29.3
pulsar-client[avro]==3.6.0
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydispatch==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
